backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.1
